Adds an app logger which writes to platforms app data location
This is built on slog, so it can be extended and configured more in the
future. I chose slog as it is native and good enough for an interactive
cli running on an end user machine.

Logs for the Buildkite MCP server are stored in the platform-specific directory:

* macOS: ~/Library/Application Support/buildkite-mcp-server/logs/
* Windows: %APPDATA%\buildkite-mcp-server\logs\
* Linux: ~/.config/buildkite-mcp-server/logs/

The name of the log file is generated from the template `{{hostname}}_{{username}}_{{timestamp:2006-01-02}}_pid{{pid}}.log` so the files are unique for different mcp servers.